Fix ARM-32 compile warning

+++ parser/parse_content.c
@@ -268,7 +268,7 @@ char* parse_content_length( char* buffer, char* end, int* length)
 	return p;
 error:
 	LM_ERR("parse error at pos %ld, dec-char: %d, start/p/end: %p/%p/%p\n",
-	       p - buffer, p < end && (end-buffer) ? *p:-1, buffer, p, end);
+	       (long)(p - buffer), p < end && (end-buffer) ? *p:-1, buffer, p, end);
 	return NULL;
 }
